+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Latest:
  -
  -
  -
- -
+ - `siac skynet upload` now supports uploading directories. All files are
+   uploaded individually and result in separate skylinks.
  -
  -
  - No user-agent needed for Skylink downloads